requirements.txt
# 基础依赖
# Python版本要求: 3.10+（请确保您的Python版本 >= 3.10）

# 文档解析
python-docx>=0.8.11
PyPDF2>=3.0.1
pdfplumber>=0.10.3
pandas>=2.0.0
openpyxl>=3.1.2
python-pptx>=0.6.21
Pillow>=10.0.0
pytesseract>=0.3.10

# 板式解析（可选，核心功能不需要）
# layoutparser>=0.3.4
# torch>=2.0.0
# torchvision>=0.15.0
# detectron2>=0.6

# NLP和向量化
transformers>=4.35.0
sentence-transformers>=2.2.2
jieba>=0.42.1

# 向量检索
faiss-cpu>=1.7.4
numpy>=1.24.0

# API框架
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0

# 其他工具
tqdm>=4.66.0
loguru>=0.7.2

